requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
pymupdf==1.24.10
Pillow==11.1.0
openai==1.51.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.20
pytest==8.3.3
pytest-asyncio==0.24.0
